+ DB clean funcs moved to lib
